requirements.txt
certifi==2020.12.5
chardet==4.0.0
fire==0.4.0
idna==2.10
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
termcolor==1.1.0
urllib3==1.26.4
